move ValidateGeoJSONFeatureCollection
